device-api-e2e/pyproject.toml
[tool.poetry]
name = "device_api_e2e"
version = "0.1.0"
description = ""
authors = ["aevtikheev <aevtikheev@ozon.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
pytest = "^7.1.2"
PyHamcrest = "^2.0.3"
grpcio-tools = "^1.48.0"
allure-python-commons = "^2.9.45"
allure-pytest = "^2.9.45"
SQLAlchemy = "^1.4.40"
psycopg2-binary = "^2.9.3"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
